requirements_airflow.txt
# --- Airflow, Experimentation & Advanced Tooling ---

# Orchestration
apache-airflow>=2.8.0
schedule>=1.2.0

# Data Versioning and Cloud
dvc>=3.0.0
dvc-s3>=3.0.0
boto3>=1.35.39

# Development and Testing
pytest>=8.3.3
pytest-cov>=5.0.0

# Jupyter (for EDA notebooks)
jupyter>=1.0.0
ipykernel>=6.29.5
